Bridge button for IBC assets (#114)
