+++ src/components/general/FinderLink.tsx
@@ -4,7 +4,7 @@ import classNames from "classnames"
 import { truncate } from "@terra.kitchen/utils"
 import { FINDER } from "config/constants"
 import { useNetworkName } from "data/wallet"
-import ExternalLink from "./ExternalLink"
+import { ExternalLink } from "./External"
 import styles from "./FinderLink.module.scss"
 
 interface Props extends HTMLAttributes<HTMLAnchorElement> {
