+++ src/components/general/index.ts
@@ -1,7 +1,7 @@
 export { default as Button } from "./Button"
 export { default as Copy } from "./Copy"
-export { default as ExternalLink } from "./ExternalLink"
-export * from "./ExternalLink"
+export { ExternalLink, ExternalIconLink } from "./External"
+export * from "./External"
 export { default as FinderLink } from "./FinderLink"
 export * from "./FinderLink"
 export * from "./Internal"
