+++ src/pages/wallet/AssetActions.tsx
@@ -5,6 +5,7 @@ import MonetizationOnOutlinedIcon from "@mui/icons-material/MonetizationOnOutlin
 import ShortcutOutlinedIcon from "@mui/icons-material/ShortcutOutlined"
 import RestartAltIcon from "@mui/icons-material/RestartAlt"
 import ArrowForwardIosIcon from "@mui/icons-material/ArrowForwardIos"
+import OpenInNewIcon from "@mui/icons-material/OpenInNew"
 import { isDenomTerraNative } from "@terra.kitchen/utils"
 import { has } from "utils/num"
 import { useIsClassic } from "data/query"
@@ -12,7 +13,11 @@ import { useNetworkName } from "data/wallet"
 import { useIsWalletEmpty } from "data/queries/bank"
 import { useCW20Pairs } from "data/Terra/TerraAssets"
 import { useTFMTokens } from "data/external/tfm"
-import { InternalButton, InternalLink } from "components/general"
+import {
+  InternalButton,
+  InternalLink,
+  ExternalIconLink,
+} from "components/general"
 import { ExtraActions } from "components/layout"
 import { ModalButton } from "components/feedback"
 import { ListGroup } from "components/display"
@@ -56,6 +61,15 @@ const AssetActions = ({ token, symbol, balance }: Props) => {
         >
           <ListGroup groups={buyList} />
         </ModalButton>
+      )}
+
+      {!isClassic && token.startsWith("ibc/") && (
+        <ExternalIconLink
+          icon={<OpenInNewIcon style={{ fontSize: 18 }} />}
+          href={`https://bridge.terra.money`}
+        >
+          {t("Bridge")}
+        </ExternalIconLink>
       )}
 
       <InternalLink
